% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/4-analysis-adjust-incidence.R
\name{adjust_incidence_1}
\alias{adjust_incidence_1}
\title{Adjust incidence 1}
\usage{
adjust_incidence_1(df, ...)
}
\arguments{
\item{df}{A data frame, the routine database dataframe, with columns conf, susp, test}

\item{...}{Group by columns}
}
\value{
Dataframe
}
\description{
Adjust incidence 1
}
\author{
Chunzhe ZHANG
}
