% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/4-analysis-adjust-incidence.r
\name{adjust_incidence_2}
\alias{adjust_incidence_2}
\title{Adjust incidence 2}
\usage{
adjust_incidence_2(df, reprat_df)
}
\arguments{
\item{df}{A data frame, calculated by adjust_incidence_1, must contain n1}

\item{reprat_df}{A data frame, containing the report rate of ad}
}
\value{
Dataframe
}
\description{
Adjust incidence 2
}
