% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/4-analysis-adjust-incidence.r
\name{adjust_incidence_3}
\alias{adjust_incidence_3}
\title{Adjust incidence 3}
\usage{
adjust_incidence_3(df, treatment_seeking_df)
}
\arguments{
\item{df}{A data frame, calculated by adjust_incidence_2, must contain n2}

\item{treatment_seeking_df}{A data frame
containing the treatment seeking rate}
}
\value{
Dataframe
}
\description{
Adjust incidence 3
}
