% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/6-report-2-plot-component.r
\name{align_plots}
\alias{align_plots}
\title{align plots
Align plots in a grid, with the same width and height in the plot area.
Use methods from the cowplot package.}
\usage{
align_plots(..., ncol = 1, align = "v", axis = "l")
}
\arguments{
\item{...}{plots to align}

\item{ncol}{number of columns, default is 1}

\item{align}{how to align, default is "v" (vertical)}

\item{axis}{axis to align, default is "l" (left)}
}
\value{
a list of plots
}
\description{
align plots
Align plots in a grid, with the same width and height in the plot area.
Use methods from the cowplot package.
}
