% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/3-dm-1-importation-shapefile.r
\name{bind_who_gis_adm2_shp}
\alias{bind_who_gis_adm2_shp}
\title{Generate ADM2 shapefiles from WHO GIS Hub}
\usage{
bind_who_gis_adm2_shp(
  shp_regional_afro_adm2,
  shp_regional_amro_adm2,
  shp_regional_emro_adm2,
  shp_regional_euro_adm2,
  shp_regional_searo_adm2,
  shp_regional_wpro_adm2,
  shp_global_adm2
)
}
\arguments{
\item{shp_regional_afro_adm2}{Path to the shapefile of the WHO African Region}

\item{shp_regional_amro_adm2}{Path to the shapefile
of the WHO American Region}

\item{shp_regional_emro_adm2}{Path to the shapefile
of the WHO Eastern Mediterranean Region}

\item{shp_regional_euro_adm2}{Path to the shapefile
of the WHO European Region}

\item{shp_regional_searo_adm2}{Path to the shapefile
of the WHO South-East Asia Region}

\item{shp_regional_wpro_adm2}{Path to the shapefile
of the WHO Western Pacific Region}

\item{shp_global_adm2}{Path to the shapefile
of the WHO Global Region}
}
\value{
A shapefile of the WHO Global Region
}
\description{
Generate ADM2 shapefiles from WHO GIS Hub
}
