% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/3-dm-2-cleaning-fuzzy-matching.r
\name{check_laps}
\alias{check_laps}
\title{Check laps}
\usage{
check_laps(df)
}
\arguments{
\item{df}{data frame generated by fuzzy_match}
}
\value{
data frame
}
\description{
Check laps
}
