% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/7-color.R
\name{complementary_color}
\alias{complementary_color}
\title{compute complementary color}
\usage{
complementary_color(color)
}
\arguments{
\item{color}{Color color}
}
\value{
complementary color
}
\description{
compute complementary color
}
