% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/0-config.R
\name{config_data_folder}
\alias{config_data_folder}
\title{config data folder}
\usage{
config_data_folder(
  root = NULL,
  root_input,
  path_raster,
  path_raster_map,
  path_raster_ihme,
  path_raster_rainfall,
  path_routine,
  path_intervention
)
}
\arguments{
\item{root}{path of the root file, could be null}

\item{root_input}{path of the input root, could be null}

\item{path_raster}{path of the raster root, could be null}

\item{path_raster_map}{path of the MAP estimation, could not be null}

\item{path_raster_ihme}{path of the IHME estimation, could not be null}

\item{path_routine}{path of the routine root, contains the routine data}

\item{path_intervention}{path of the intervention root,
contains the intervention data}

\item{path_rainfall}{path of the rainfall data, could not be null}
}
\value{
a list of data folder
}
\description{
config data folder for SNT analysis
}
