% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/0-config.R
\name{config_db}
\alias{config_db}
\title{Config database}
\usage{
config_db(db_name = NULL, db_user = NULL, db_pass = NULL)
}
\arguments{
\item{db_name}{name of database, example malaria}

\item{db_user}{username of the database, for connection}

\item{db_pass}{password of the database, for connection}
}
\value{
a list contain the db info
}
\description{
Config database
}
