% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/0-config.R
\name{config_dhs}
\alias{config_dhs}
\title{config DHS folders}
\usage{
config_dhs(path_input = NULL, path_dhs = NULL)
}
\arguments{
\item{path_input}{path of the input root, could be NULL}

\item{path_dhs}{path of the DHS root}
}
\value{
a list of DHS folders
}
\description{
config DHS folders for SNT analysis
}
