% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/0-config.R
\name{config_input}
\alias{config_input}
\title{config input folder}
\usage{
config_input(
  path_input,
  path_raster,
  path_raster_map,
  path_raster_ihme,
  path_raster_rainfall,
  path_country,
  path_routine,
  path_intervention,
  path_dhs
)
}
\arguments{
\item{path_input}{path of the input root, could be NULL}

\item{path_raster}{path of the raster root, could be NULL}

\item{path_raster_map}{path of the MAP estimation}

\item{path_raster_ihme}{path of the IHME estimation}

\item{path_country}{path of the country root, contains the country submitted
data}

\item{path_routine}{path of the routine root, contains the routine data}

\item{path_intervention}{path of the intervention}

\item{path_dhs}{path of the DHS data}

\item{path_rainfall}{path of the rainfall data}
}
\value{
a list of input folder
}
\description{
config input folder for SNT analysis
}
