% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/0-config.R
\name{config_parallel}
\alias{config_parallel}
\title{Configuration for parallel execution}
\usage{
config_parallel()
}
\value{
a list contain the parallel execution info
}
\description{
Configuration for parallel execution
}
