% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/0-config.R
\name{config_raster}
\alias{config_raster}
\title{config raster folders}
\usage{
config_raster(
  path_raster,
  path_raster_map,
  path_raster_ihme,
  path_raster_rainfall
)
}
\arguments{
\item{path_raster}{path of the raster root, could be NULL}

\item{path_raster_map}{path of the MAP estimation, could not be null}

\item{path_raster_ihme}{path of the IHME estimation, could not be NULL}

\item{path_raster_rainfall}{path of the rainfall data, could not be NULL}
}
\value{
a list of raster folders
}
\description{
config raster folders for SNT analysis
}
