% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/4-analysis-consistency.r
\name{consistency_check}
\alias{consistency_check}
\title{Consistency check}
\usage{
consistency_check(df, index_a, index_b)
}
\arguments{
\item{df}{dataframe}

\item{index_a}{smaller index for comparison}

\item{index_b}{larger index for comparison}
}
\value{
plot
}
\description{
Consistency check
}
\author{
Chunzhe ZHANG
}
