% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/20-dev-tools.R
\name{dev_update}
\alias{dev_update}
\title{Devtools}
\usage{
dev_update()
}
\description{
This command is a combination of several commands to update the package
It will run the following commands:
\itemize{
\item \code{update_database()}
\item \code{devtools::document()}
\item \code{devtools::load_all()}
\item \code{devtools::test()}
\item \code{devtools::check()}
\item \code{pkgdown::build_site()}
}
}
\seealso{
update_database
}
