% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/3-dm-1-importation-shapefile.r
\name{extract_adm1_adm2_from_who_gis_adm2_shp}
\alias{extract_adm1_adm2_from_who_gis_adm2_shp}
\title{extract adm1 adm2 list from WHO GIS hub adm2 shapefile}
\usage{
extract_adm1_adm2_from_who_gis_adm2_shp(shp_global_adm2)
}
\arguments{
\item{shp_global_adm2}{Path to the shapefile of the WHO Global Region}
}
\value{
A list of adm1 adm2 with GUID
}
\description{
extract adm1 adm2 list from WHO GIS hub adm2 shapefile
}
