% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/2-shapefile.R
\name{extract_country_from_who_gis_adm2_shp}
\alias{extract_country_from_who_gis_adm2_shp}
\title{extract a country from WHO GIS hub adm2 shapefile}
\usage{
extract_country_from_who_gis_adm2_shp(shp_global_adm2, iso3, save_path)
}
\arguments{
\item{shp_global_adm2}{Path to the shapefile of the WHO Global Region}

\item{iso3}{ISO3 code Country to extract}

\item{save_path}{Path to save the shapefile}
}
\description{
extract a country from WHO GIS hub adm2 shapefile
}
