% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/2-preprocess-extract-who-gis.R
\name{extract_shp_from_gishub}
\alias{extract_shp_from_gishub}
\title{Extract shapefile for a country}
\usage{
extract_shp_from_gishub(p_gishub, iso3, p_out)
}
\description{
Extract shapefile for a country
p_gishub: path to the shapefile from the GISHub
iso3: ISO3 code of the country
p_out: path to the output shapefile
}
\examples{
extract_shp_from_gishub(
    p_gishub = p_shapefile_adm2,
    iso3 = "COG",
    p_out = p_congo_shp
)
congo_shp <- st_read(p_congo_shp)
plot(congo_shp)
}
\author{
Chunzhe ZHANG
}
