% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-math.R
\name{faster_mean}
\alias{faster_mean}
\title{Faster Mean}
\usage{
faster_mean(data)
}
\arguments{
\item{data}{a Combination of numeric values}
}
\value{
a number
}
\description{
Detailed the reasons please refer to this
\href{https://stackoverflow.com/a/18604487/886198}{answer}
}
\examples{
faster_mean(c(1,2,3))
}
\author{
Spencer Zhang (sepmein@gmail.com)
}
