% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/2-preprocess-outlier.R
\name{find_outlier}
\alias{find_outlier}
\title{Find outlier}
\usage{
find_outlier(
  df,
  columns = c("susp", "test_rdt", "test_mic", "test_rdt_lab", "test", "abn_mic",
    "abn_rdt", "conf_rdt", "maltreat_u5", "maltreat_ov5", "maltreat", "maldth"),
  alpha = 0.999,
  both_sides = FALSE,
  select_rows = c("ID", "adm1", "adm2", "adm3", "hfca", "hfname", "hf", "year", "month",
    "yearmon")
)
}
\arguments{
\item{df}{dataframe}

\item{columns}{columns to be checked}

\item{alpha}{alpha value for quantile}

\item{both_sides}{whether to check both sides of the distribution}

\item{select_rows}{columns to be selected}
}
\value{
dataframe
}
\description{
Find outlier in the dataset
}
\author{
Chunzhe ZHANG
}
