% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/3-dm-2-cleaning-outliers.r
\name{find_outliers}
\alias{find_outliers}
\title{Find outlier}
\usage{
find_outliers(df, alpha = 0.99999, both_sides = FALSE, ...)
}
\arguments{
\item{df}{dataframe}

\item{alpha}{alpha value for quantile}

\item{both_sides}{whether to check both sides of the distribution}

\item{columns}{columns to be checked}

\item{select_rows}{columns to be selected}
}
\value{
dataframe
}
\description{
Find outlier in the dataset
}
\author{
Chunzhe ZHANG
}
