% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/2-preprocess-fuzzy-match.R
\name{fuzzy_match}
\alias{fuzzy_match}
\title{Fuzzy match}
\usage{
fuzzy_match(df, col, threshold = 2, method = "osa")
}
\arguments{
\item{df}{Dataframe}

\item{col}{Column to match}

\item{threshold}{Threshold for fuzzy match}

\item{method}{Method for fuzzy match}
}
\value{
Dataframe
}
\description{
Fuzzy match function, check this answer for more details:
https://stackoverflow.com/a/41560524/886198
}
