% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/3-dm-2-cleaning-fuzzy-matching.r
\name{fuzzy_match_enhanced}
\alias{fuzzy_match_enhanced}
\title{Fuzzy match result
Identify and remove duplicate reports}
\usage{
fuzzy_match_enhanced(df, col, threshold = 1.1, method = "lv", ...)
}
\arguments{
\item{df}{Dataframe
df: dataframe to be checked for duplicates}

\item{col}{column to be used for fuzzy matching}

\item{threshold}{minimum match level for fuzzy matching, default is 1.1}

\item{method}{fuzzy matching method, default is "lv"}

\item{...}{group_by columns to get the report status and duration for each group}
}
\value{
Dataframe
}
\description{
Fuzzy match result
Identify and remove duplicate reports
}
