% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/0-config.R
\name{generate_default_gitignore}
\alias{generate_default_gitignore}
\title{Generate Default .gitignore}
\usage{
generate_default_gitignore()
}
\description{
Write .gitignore if not exists
}
\seealso{
https://github.com/johnaponte/repana
}
\keyword{file}
\keyword{git,}
\keyword{gitignore,}
