% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/2-ra-path.r
\name{get_os}
\alias{get_os}
\title{Get OS}
\usage{
get_os()
}
\value{
string of operating system
}
\description{
Get Operating System from R
thanks to will at  https://www.r-bloggers.com/identifying-the-os-from-r/
}
