% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/0-config.R
\name{get_path}
\alias{get_path}
\title{Get path}
\usage{
get_path(type, path)
}
\arguments{
\item{type}{type of path, could be 'data',}

\item{path}{path of the file}
}
\value{
return the path
}
\description{
Get path from config.yml and from user defined file
}
\examples{
get_path("data", "test.csv")
}
