% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/7-routine-report-rate.R
\name{get_report_duration}
\alias{get_report_duration}
\title{Calculate the report duration
Get the min max report year and month for each group based on the report status
This code filters the data by the reported column to only include data
that is reported, then groups the data by the variables specified in the
function call. It then calculates the min and max year and month for each
group, and then filters out any groups that have NA for the min year.}
\usage{
get_report_duration(df, ...)
}
\arguments{
\item{df}{A data frame with report status generated by report_status()}

\item{...}{Columns to group by}
}
\value{
Dataframe
}
\description{
Calculate the report duration
Get the min max report year and month for each group based on the report status
This code filters the data by the reported column to only include data
that is reported, then groups the data by the variables specified in the
function call. It then calculates the min and max year and month for each
group, and then filters out any groups that have NA for the min year.
}
