% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/3-dm-1-importation-smart-read.r
\name{smart_read_excel_by_year}
\alias{smart_read_excel_by_year}
\title{Read excel file by year}
\usage{
smart_read_excel_by_year(smart_path, skip = 0, clean = TRUE)
}
\arguments{
\item{smart_path}{A string, the path of the file. The path should
contain the pattern of \\emph{yyyy\}. Otherwise, the function will return
an error.}

\item{skip}{An integer, the number of rows to skip.}

\item{clean}{A boolean, whether to clean the data.
\itemize{
\item \code{TRUE}(the default) will clean the data. The cleaning process will use
internal database, \code{snt::routine_rename} to rename and replace the data.
\item \code{FALSE} will not clean the data.
}}
}
\description{
The function will read all excel file follows the
pattern of \\\emph{yyyy\\} in the path.
The function will return a nested tibble.
}
\examples{
\dontrun{
# read excel file by year
smart_read_excel_by_year(
  "C:/Users/username/Desktop/Bo District/Bo District *yyyy*.xlsx"
)
}
}
\seealso{
routine_rename
}
