% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import.R
\name{import_replace}
\alias{import_replace}
\title{A helper function to import routine database}
\usage{
import_replace(df, replace_database, snt_country, column, by = FALSE)
}
\arguments{
\item{df}{Target dataframe to be replaced}

\item{replace_database}{Db which stored the corrected information, usually
snt internal database could be used.}

\item{snt_country}{Target snt country, as there is multiple country replacement
store inside the replace db. So designate a database is necessary}

\item{column}{Target column in the target dataframe to be replaced}
}
\description{
A helper function to import routine database
}
\author{
Chunzhe ZHANG
}
