% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/5-data.R
\docType{data}
\name{import_routine_rename}
\alias{import_routine_rename}
\title{Database for Rename column names in Routine Data}
\format{
A tibble with BDI:
\describe{
\item{country}{country code, ISO3 code}
\item{from}{column name to be formatted, string}
\item{to}{column name to be formatted, string}
...
}
}
\usage{
import_routine_rename
}
\description{
A dataset containing the country code
original column names and the formatted column names
import_routine_rename.
}
\keyword{datasets}
