% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Data.R
\docType{data}
\name{import_routine_replace}
\alias{import_routine_replace}
\title{Database for Replace column values in Routine Data}
\format{
A tibble with BDI:
\describe{
\item{country}{country code, ISO3 code}
\item{from}{value to be formatted, string}
\item{to}{value to be formatted, string}
...
}
}
\usage{
import_routine_replace
}
\description{
A dataset containing the country code, original value, and the formatted value
import_routine_replace.
}
\keyword{datasets}
