% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Data.R
\docType{data}
\name{import_routine_replace_hfname}
\alias{import_routine_replace_hfname}
\title{Database for Replace health facilities in Routine Data}
\format{
A tibble with BDI:
\describe{
\item{country}{country code, ISO3 code}
\item{district}{district name, string}
\item{from}{hf name to be formatted, string}
\item{to}{hf name to be formatted, string}
...
}
}
\usage{
import_routine_replace_hfname
}
\description{
A dataset containing the country code, district name, original health facilities name, and the formatted health facilities name
import_routine_replace_hfname.
}
\keyword{datasets}
