% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Data.R
\docType{data}
\name{import_routine_set_cluster}
\alias{import_routine_set_cluster}
\title{Database for Set Adm2 Cluster in Routine Data}
\format{
A tibble with BDI:
\describe{
\item{country}{country code, ISO3 code}
\item{adm2}{adm2 name, string}
\item{cluster}{cluster, int}
...
}
}
\usage{
import_routine_set_cluster
}
\description{
A dataset containing the country code, adm2, and the cluster
import_routine_set_cluster.
}
\keyword{datasets}
