% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/2-preprocess-dups.R
\name{merge_dups}
\alias{merge_dups}
\title{Merge duplicates in a data frame}
\usage{
merge_dups(df, ...)
}
\arguments{
\item{df}{Dataframe}

\item{...}{Column to check for duplicates}
}
\value{
Dataframe
}
\description{
Merge duplicates in a data frame
}
