% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-tibble.R
\name{odd_or_even}
\alias{odd_or_even}
\title{Get odd or even rows from a tibble}
\usage{
odd_or_even(df, odd = TRUE)
}
\arguments{
\item{df}{tibble}

\item{odd}{boolean, if TRUE then get odd rows, if FALSE then get even rows}
}
\value{
tibble
}
\description{
Get odd or even rows from a tibble
}
