% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/4-analysis-report-rate.r
\name{plot_adm_report_status}
\alias{plot_adm_report_status}
\title{Plot Report Status by district month}
\usage{
plot_adm_report_status(df, ..., index = c("conf", "test", "susp"))
}
\arguments{
\item{df}{A data frame}

\item{...}{Group by columns}

\item{index}{Index columns}
}
\value{
Dataframe
}
\description{
Plot Report Status by district month
}
