% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/4-analysis-report-rate.R
\name{plot_hf_report_status}
\alias{plot_hf_report_status}
\title{Plot Report Status}
\usage{
plot_hf_report_status(df, ...)
}
\arguments{
\item{df}{A data frame}

\item{...}{Columns to group by}
}
\value{
Dataframe
}
\description{
Plot Report Status
}
