% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/4-analysis-report-rate.R
\name{plot_indicator_report_status}
\alias{plot_indicator_report_status}
\title{Plot Report Status by indicators}
\usage{
plot_indicator_report_status(df, ...)
}
\arguments{
\item{df}{A data frame}

\item{...}{indicator columns}
}
\value{
Dataframe
}
\description{
Plot Report Status by indicators
}
\author{
Chunzhe ZHANG
}
