% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/4-analysis-adjust-incidence.R
\name{plot_map_by_year}
\alias{plot_map_by_year}
\title{Plot incidences}
\usage{
plot_map_by_year(
  df,
  col,
  breaks = c(0, 250, 350, 450, 20000),
  palette = hcl.colors(4, "Blue-Red")
)
}
\arguments{
\item{df}{A data frame}

\item{col}{Column name to plot}

\item{breaks}{Breaks for the color scale}

\item{palette}{Color palette}
}
\description{
Plot incidences
}
