% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/3-dm-2-cleaning-outliers.r
\name{plot_outliers}
\alias{plot_outliers}
\title{Plot outliers}
\usage{
plot_outliers(df, ..., save_to = NULL)
}
\arguments{
\item{df}{dataframe}

\item{...}{columns to be plotted}

\item{save_to}{path to save the plot}
}
\value{
plot
}
\description{
Plot outliers
}
\author{
Chunzhe ZHANG
}
