% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/2-preprocess-missing.R
\name{remove_and_export_missing}
\alias{remove_and_export_missing}
\title{Remove and export missing data}
\usage{
remove_and_export_missing(df, path, ...)
}
\arguments{
\item{df}{Dataframe}

\item{path}{Path to export missing data}

\item{...}{Column to check for missing data}
}
\value{
Dataframe
}
\description{
Remove and export missing data
}
