% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/1-preprocess-replace-double-space.R
\name{replace_double_space}
\alias{replace_double_space}
\title{Replace double space
replace two or more spaces with one space}
\usage{
replace_double_space(df, column)
}
\arguments{
\item{df}{dataframe}

\item{column}{column to replace}
}
\value{
dataframe
}
\description{
Replace double space
replace two or more spaces with one space
}
\examples{
df <- data.frame(
    col = c("a  b", "b  c", "c  d")
)
replace_double_space(df, col)
}
