% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/5-snt-report-rate.r
\name{report_rate}
\alias{report_rate}
\title{Calculate the reporting rate for each indicator}
\usage{
report_rate(df, ...)
}
\arguments{
\item{df}{A data frame}

\item{...}{Columns to calculate the reporting rate, excluding year and month}
}
\value{
Dataframe
}
\description{
Calculate the reporting rate for each indicator
}
