% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/5-snt-report-rate.r
\name{report_status}
\alias{report_status}
\title{Calculate the report status}
\usage{
report_status(df, ...)
}
\arguments{
\item{df}{A data frame}

\item{...}{Columns to group by}
}
\value{
Dataframe
}
\description{
If any indicator is reported, the report status is 'Y', otherwise 'N'
grouped by the provided columns
}
