% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/3-dm-3-integration-renaming-header.r
\name{routine_rename}
\alias{routine_rename}
\title{Rename Columns in Routine data}
\usage{
routine_rename(df)
}
\arguments{
\item{df}{dataframe to be renamed}
}
\description{
Rename columns in routine data
}
