% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/3-dm-2-cleaning-replace-strings.r
\name{routine_replace}
\alias{routine_replace}
\title{Replace routine database}
\usage{
routine_replace(df)
}
\arguments{
\item{df}{Target dataframe to be replaced}
}
\description{
Replace routine database
}
