% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/6-report-2-plot-component.r
\name{scale_numbered}
\alias{scale_numbered}
\title{Numbered axis}
\usage{
scale_numbered(axis = "x")
}
\arguments{
\item{axis}{x or y axis to be reformatted, default is x}
}
\description{
reformat axis to be numbered, examples is format from
1e6 to 1,000,000
}
