% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/0-config.R
\name{setup}
\alias{setup}
\title{Setup a folder for analysis}
\usage{
setup(path)
}
\arguments{
\item{path}{Path to where you want the folder will be}
}
\description{
Provide a path, snt will setup a whole folder settings for you.
}
