% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util-file-reader.R
\name{smart_get_file_list_by_year}
\alias{smart_get_file_list_by_year}
\title{Smart file reader}
\usage{
smart_get_file_list_by_year(smart_path)
}
\arguments{
\item{smart_path}{A string, the path of the file.}
}
\value{
A list of file path.
}
\description{
This function is a smart file reader,
it will detect the file path and return the file list.
}
\examples{
\dontrun{
# get file list by year
smart_get_file_list_by_year(
"C:/Users/username/Desktop/Bo District/Bo District *yyyy*.xlsx")
}
}
