% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/snt-package.R
\docType{package}
\name{snt-package}
\alias{snt}
\alias{snt-package}
\title{snt: Malaria country analysis}
\description{
What the package does (one paragraph).
}
\author{
\strong{Maintainer}: Chunzhe Zhang \email{zhangc@who.int} (\href{https://orcid.org/YOUR-ORCID-ID}{ORCID})

}
\keyword{internal}
