% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/3-dm-2-cleaning-str.r
\name{str_slice_to_end}
\alias{str_slice_to_end}
\title{String Util Slice a pattern and append it to the end}
\usage{
str_slice_to_end(df, column, pattern)
}
\arguments{
\item{pattern}{target pattern to match}

\item{str}{target string}

\item{concatenate}{optional concatenate string, default to a space}
}
\value{
string
}
\description{
String Util Slice a pattern and append it to the end
}
\examples{
'General Hospital Awa Onna'
str_slice_to_end('General Hospital Awa Onna', 'General Hospital')
'Awa Onna General Hospital'
}
