% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/3-dm-2-cleaning-str.r
\name{str_trim_space}
\alias{str_trim_space}
\title{Replace double space
replace two or more spaces with one space}
\usage{
str_trim_space(df, column)
}
\arguments{
\item{df}{dataframe}

\item{column}{column to replace}
}
\value{
dataframe
}
\description{
Replace double space
replace two or more spaces with one space
}
\examples{
df <- data.frame(
  col = c('a  b', 'b  c', 'c  d')
)
str_trim_space(df, col)
}
