% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/7-calc.r
\name{sum_maltreat_ov5_by_sex}
\alias{sum_maltreat_ov5_by_sex}
\title{Sum maltreat over 5 years old by female and male}
\usage{
sum_maltreat_ov5_by_sex(
  df,
  female_col = "maltreat_ov5_f",
  male_col = "maltreat_ov5_m"
)
}
\arguments{
\item{df}{dataframe}
}
\value{
dataframe
}
\description{
Sum maltreat over 5 years old by female and male
}
\examples{
sum_maltreat_ov5_by_sex(df)
}
