% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/5-snt-maltreat.r
\name{sum_maltreat_u5_by_sex}
\alias{sum_maltreat_u5_by_sex}
\title{Sum maltreat under 5 years old by female and male}
\usage{
sum_maltreat_u5_by_sex(
  df,
  female_col = "maltreat_u5_f",
  male_col = "maltreat_u5_m"
)
}
\arguments{
\item{df}{dataframe}
}
\value{
dataframe
}
\description{
Sum maltreat under 5 years old by female and male
}
\examples{
sum_maltreat_u5_by_sex(df)
}
