% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/3-dm-4-transformation-aggregation.r
\name{sum_without_na}
\alias{sum_without_na}
\title{Sum without NA}
\usage{
sum_without_na(data)
}
\arguments{
\item{data}{a Combination of numeric values}
}
\value{
a number
}
\description{
Sum without NA
}
\examples{
sum_without_na(c(1, 2, 3, NA))
}
\author{
Spencer Zhang (sepmein@gmail.com)
}
