% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/2-preprocess-renaming.R
\name{to_lower}
\alias{to_lower}
\title{Renaming the dataset to lower case}
\usage{
to_lower(df)
}
\arguments{
\item{df}{dataframe to be renamed}
}
\description{
Renaming the dataset to lower case
}
\examples{
\dontrun{
# rename the dataframe to lower case
df <- tibble::tibble(
    A = 1:3,
    B = 4:6,
    C = 7:9
)
df_rename <- snt::to_lower(df)
}
# result
# A tibble: 3 x 3
#   a     b     c
#   <int> <int> <int>
# 1     1     4     7
# 2     2     5     8
# 3     3     6     9
}
