% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/2-ra-config.r
\name{type_of_path}
\alias{type_of_path}
\title{Check type of path}
\usage{
type_of_path(path)
}
\arguments{
\item{path}{String path}
}
\description{
check type of path. check the existence of the dir or file
if they exist then return the type of the path
else return false
}
