% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/4-dev-update-database.R
\name{update_database}
\alias{update_database}
\title{Devtools : Update SNT databases}
\usage{
update_database()
}
\description{
When importing data, snt often requires a lot of renaming
those db were created for reusable renaming process
}
\examples{
# edit db in inst/extdata, then run
update_database()

}
