% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/0-config.R
\name{use_relative_or_absolute}
\alias{use_relative_or_absolute}
\title{Check to use relative or absolute path}
\usage{
use_relative_or_absolute(root, relative)
}
\arguments{
\item{root}{String, root path}

\item{relative}{String, relative path}
}
\value{
path
}
\description{
Combine root and relative check if they exist in path
Check relative path exists
if it is true then use the combined path
if it is false then use the relative path
}
