% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/2-preprocess-renaming.R
\name{adm_rename}
\alias{adm_rename}
\title{rename the administrative zone names}
\usage{
adm_rename(df)
}
\arguments{
\item{df}{dataframe to be renamed}
}
\description{
Detect if adm0, adm1, adm2, adm3 are in the dataframe, if they are
use the standard names in the snt internal database to rename them
}
