% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/3-dm-2-cleaning-duplicate.r
\name{check_dups}
\alias{check_dups}
\title{Check for duplicates in a data frame}
\usage{
check_dups(df, ...)
}
\arguments{
\item{df}{Dataframe}

\item{...}{Column to check for duplicates}
}
\value{
Dataframe
}
\description{
Check for duplicates in a data frame
}
