% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/0-config.R
\name{config}
\alias{config}
\title{Central Configuration for SNT analysis}
\usage{
config(
  country,
  root = NULL,
  root_input = NULL,
  root_output = NULL,
  root_raster = NULL,
  raster_map_input = NULL,
  ihme_folder = NULL,
  rainfall_folder = NULL,
  rainfall_output = NULL,
  shapefile = NULL,
  db_name = NULL,
  db_user = NULL,
  db_pass = NULL
)
}
\arguments{
\item{country}{ISO3 code for a country, e.g. nigeria should be NGA}

\item{root}{path of the root file, could be NULL
if set, will be combined with other root folders}

\item{root_input}{path of the input root, could be NULL}

\item{root_output}{path of the output root. If set, will be combined with}

\item{shapefile}{}
}
\description{
Central Configuration settings for SNT analysis.
This function could set the following configurations:
\enumerate{
\item the target subnational analysis country
\item folder structure
All the folder structure could be set separately,
however it will be nice if they are all under one root folder.
Separate setting
\item root
\item postgresql db connection
}
}
\examples{
config(
  country = "NGA",
  root = "/Users/sepmein/snt/NGA",
  root_input = "input",
  root_output =
  )
}
