% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/2-ra-config.r
\name{config_set_db}
\alias{config_set_db}
\alias{config_get_db}
\alias{config_parallel}
\alias{config_get_country}
\title{Config database}
\usage{
config_set_db(db = "SQLite", db_name = NULL, db_user = NULL, db_pass = NULL)

config_get_db()

config_parallel()

config_get_country()
}
\arguments{
\item{db_name}{name of database, example malaria}

\item{db_user}{username of the database, for connection}

\item{db_pass}{password of the database, for connection}
}
\value{
a list contain the db info

db connection

a list contain the parallel execution info

country
}
\description{
Get the country from the yaml configuration file
TODO: format country name to ISO3 country code
}
\examples{
config_get_country()
}
