% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/0-config.R
\name{config_country}
\alias{config_country}
\title{config country folders}
\usage{
config_country(country)

config_country(country)
}
\arguments{
\item{country}{The country name}

\item{path_country}{path of the country root}

\item{path_routine}{path of the routine, contains the routine data}

\item{path_intervention}{path of the intervention
contains the intervention data}
}
\value{
a list of country folders

ISO3 country code
}
\description{
config country folders for SNT analysis
}
