% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/1-preprocess-fuzzy-match.R
\name{fuzzy_match}
\alias{fuzzy_match}
\title{Fuzzy match}
\usage{
fuzzy_match(df, col, threshold = 1.1, method = "lv", ...)
}
\arguments{
\item{df}{Dataframe}

\item{col}{Column to match}

\item{threshold}{Threshold for fuzzy match}

\item{method}{Method for fuzzy match}
}
\value{
Dataframe
}
\description{
Fuzzy match function, check this answer for more details:
https://stackoverflow.com/a/41560524/886198
}
\examples{
df <- data.frame(
    col = c("a", "b", "c", "aa")
)
fuzzy_match(df, col, match_col)
fuzzy_match(df, col, match_col, threshold = 1)
}
\author{
\link{email protected}
}
