% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/0-config.R
\name{generate_default_config}
\alias{generate_default_config}
\title{Generate Default Config}
\usage{
generate_default_config()
}
\description{
Write config.yml if not exists
}
\seealso{
https://github.com/johnaponte/repana
}
