% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import.R
\name{import_dhs_data}
\alias{import_dhs_data}
\title{Import DHS data from API}
\usage{
import_dhs_data(year = 2021)
}
\arguments{
\item{year}{DHS year to be imported}
}
\value{
A tibble of DHS data
}
\description{
Import DHS data from API
}
\author{
Chunzhe ZHANG
}
