% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/4-analysis-report-rate.r
\name{plot_hf_report_status}
\alias{plot_hf_report_status}
\title{Plot Report Status}
\usage{
plot_hf_report_status(report_status)
}
\arguments{
\item{df}{A data frame}
}
\value{
Dataframe
}
\description{
Plot Report Status
}
