% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/snt-package.R
\docType{package}
\name{snt-package}
\alias{snt}
\alias{snt-package}
\title{snt: Malaria Subnational Tailoring}
\description{
This package provides functions to download, process and visualize malaria data at subnational level.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/sepmein/snt}
  \item \url{https://sepmein.github.io/snt/}
  \item Report bugs at \url{https://github.com/sepmein/snt/issues}
}

}
\author{
\strong{Maintainer}: Chunzhe Zhang \email{zhangc@who.int} (\href{https://orcid.org/YOUR-ORCID-ID}{ORCID})

}
\keyword{internal}
