% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/1-preprocess-type.R
\name{transmute_numeric}
\alias{transmute_numeric}
\title{Transmute numeric columns to the correct type}
\usage{
transmute_numeric(df)
}
\arguments{
\item{df}{A data frame}
}
\value{
A data frame
}
\description{
Transmute numeric columns to the correct type
}
